Make callback scheduling success/failure more explicit.
This CL introduces a new AsyncResult type which is yielded
from functions which may or may not schedule callbacks. Functions
of this type are changed to use AsyncResult rather than raw bools.

We also ensure that qsri callbacks are always fired, even if they
are not properly scheduled. This fixes a deadlock when VkImage
unboxing fails.

Bug: 238334576
Bug: 238333419
Test: Freeze detailed in 238334576 no longer occurs.

Change-Id: I4a3a752960d9900f674fa637f483c8b2100da068