Revert "emugl_config: fix potential uses of nullptr"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/07e2b7cb-1a89-4ec3-9053-a8cca826c2fb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7370e0ce25edd927285e544c05659017277ab19d