Revert "emugl_config: fix potential uses of nullptr"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/a80cf050-8869-4ddd-a06d-ea9a3524e298). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I671fc64dc1bcac3c22e1322945128569fb64dd88