Revert "aemu: stop requiring lz4, Googletest, and perfetto-traci..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9e543b96-27e0-45bc-95da-703f9cb34aa3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icd96bcce505d9f819151401805c3072911c0ab36