Revert "aemu: stop requiring lz4, Googletest, and perfetto-traci..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/997ea54e-2bd4-48a9-b3ac-1d929d87777e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I86c9d21257141e54dc56a611e0569643eff3e25b