Revert "aemu: stop requiring lz4, Googletest, and perfetto-traci..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1f65508c-6baf-4dcc-94e9-2c1255393ce3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I857040d4a6095910fbfb3cc7afd36f5cf8a6b457