Revert "update hw-config-defs to include userdata.useQcow2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d558179b-c4e1-40bd-973e-3e0f3f1e9816). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifea356e4134c8e40fdf603c8b21cfd2eef2af22b