Revert "update hw-config-defs to include userdata.useQcow2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/069c8e22-2069-4927-b7b7-3194362e8ed1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ife104ed730749ccfa9a1b9aeb1403ddb3e22fdba