Revert "Converge log severity in Goldfish emulator to AOSP's def..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/836565ec-73ac-494e-a523-dc1a70d00bd6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I98dd1714b6d33114a96a8ef34d869bf1ead55fb8