Revert "aemu: build SubAllocator in host-common too"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2823c34f-a41e-4f7e-93e8-d89782dc7f53). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6155425b8b07c89ee72cf9f0d928d494dafdc0a2