Apply stacked layout for automotive devices
Especially for the wide width displays such as Distant Display, it's
hard to use the emulator when all displays are located on one row.
This change locates a display on a separate row when the total width
exceeds 50% of monitor width.

Bug: 316033487
Test: launch multi display targets such as gcar_dd_x86_64,
gcar_md_x86_64 and gcar_x86_86 in main trunk

Change-Id: I7c9846985ea147ff94f2374cd44840af5ae29b5d