Allow mixing add and addFixed in HybridEntityManager
HybridEntityManager was designed under the assumption that add and
addFixed does not interleave with each other. But our Vulkan
implementation calls set_boxed_non_dispatchable_VkImage and
set_boxed_non_dispatchable_VkFence which ends up calling addFixed in the
middle of other add calls.

In the original design this would have the side effects that it might
assign a new object to an existing entry. We update the implementation
to prevent it.

Bug: 328105901
Test: HybridEntityManager.UpdateIndex
Change-Id: I81959917d18f99204a378afda15639f6f956249d