Bazel build fixes
This fixes a series of issues that make sure we can build under bazel in
C++20.

Note that we have to disable the static_assert for C++20. It turns out
that std::atomic_int didn't get initialized by default, which has
changed in C++20. This in turn makes std::atomic_int no longer default
initializable.

Change-Id: Icc15480b1a0262467a568e5c9d7ca56f31e8c119