Move msvc compat layer from emu-dev
We now create explicit header files for each compatibility piece we are
using. This will make it possible to include aemu-base inside qemu
plugins and include a bazel compat library that can be used as well.

The individual files are guarded so that there are no conflicts between
qemu 8 and aemu.

Bug: 363313973
Change-Id: Idd676c3e3da3185b22a7edbdece2e819221cf315