Rewrite resizeImpl
1. using nullptr as InputIterator is undefined
See page 955 of the following C++ 17 standard draft:
https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2017/n4659.pdf

Note that mBuffer is always nullptr at least the first time
resizeImpl is called.

2. Copy once, not twice
Due to the semantics of reallocImpl, when mBuffer should be
resized for bigger, the copy currently happens twice. One
from mBuffer to temp, and another from temp to the newly
allocated, empty/dirty mBuffer.

It is doable and more efficient to copy the data in mBuffer
only once.

3. A few formatting issues, etc
Google C++ style encourages to use {} for if block. Besides,
there seem to be a couple of coding style issues.

Test: built on Linux
Bug: N/A
Change-Id: I9d12e0df21dfbd5213d0858a93c49d73903018c8
Signed-off-by: Kyoungwon Kim <kyoungwon.kim@bytedance.com>