AlignedBuf makes sense only for trivial objects
When it creates the buffer for the objects of type T, it
calls something-like malloc, which does not call the
constructor of T. When it resizes or copies, the basically
copies the memory contents, which may lead to undefined
behavior for non-trivial objects.

Thus, the static assert should be for the entire class,
not just for the resize method.

Test: build Android Emulator
Bug: N/A
Change-Id: If9d1860ea183858a841f9b2d3be0103008fd4463
Signed-off-by: Kyoungwon Kim <kyoungwon.kim@bytedance.com>