Revert "Implement logforwarding getter for abseil logger"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/335eda04-afa3-4d49-bd34-c87344943751). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibede93cf7667ee337fbf0edab24762effee75094
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com