Revert "Reapply "Add methods to check multi display state""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e0aacde0-39a9-41af-bf13-f21dc9b322fb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I54e815282f202630c20f5ba2cf86d11e1290a9fa
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com