Add an emu feature defition: GoldfishAddressSpaceOOBAccess.
We observed that some times the guest could read/write to an address
that is not backed by RAM at the moment of access. It is likely a
wrong address from guest graphics stack. However, it is hard to find
out where.

Add an emu feature "GoldfishAddressSpaceOOBAccess" to ignore such
accesses so that the guest can continue (instead of an emulator
crash). This is for debug purpose.

Bug: 389987556

Change-Id: I157145230505283dd9c7ec4b9b5cbdaa75e00ff7