Revert "Add toString_SnapshotSkipReason utility function"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/98ae2e9b-e13c-4c1e-a6df-ffbe23a4978a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I05568dd8c1496e52a78d12764a2ec6d8d7105197
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com