Revert "Add toString_SnapshotSkipReason utility function"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/10c9893c-7499-44a1-b1de-80ff99824975). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie18e9b11c0862b1e17da87c44fd98b7e649aefe4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com