Revert "snapshot: provide interface to register vulkan apps"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/78baee7e-74f3-4aa4-8c7e-9f64d1df4056). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1007a19adb99f9769bb67f41a4e9cf0f466a588d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com