Revert "Port debug.cpp implementation from emulator"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/16c2d30b-8a30-4e57-9620-8ccbc48db80d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie62eec44c8b0a1e364541dfb7da1251eeea960a1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com