Add an entry to enable the HEADING sensor
from hardware/interfaces/sensors/aidl/android/hardware/sensors/SensorType.aidl:

A sensor of this type measures the direction in which the device is
pointing relative to true north in degrees.

This sensor was added for automotive form factors. Other devices with
a  clear forward direction might find it useful as well.

Bug: 348282766
Test: presubmit
Change-Id: I5dda343bfad974fff5c607a85853575cda09ca81
Signed-off-by: Roman Kiryanov <rkir@google.com>