feat: Add event support library
This library introduces a set of classes to simplify event handling. It
includes EventWaiter for blocking and waiting on events,
EventChangeSupport for event-driven programming, and
PropertyChangeSupport for tracking property value changes.

Additionally, it provides WithCallbacks for callback-based event
handling and ScopedEventCallback for RAII-style callback management.

Includes a set of unit tests to validate behavior.

Bug: 331949836
Change-Id: Ia51404319b300064997a54592b1936fe47a38712