refactor(base/events): Add ScopedCallbackHandle alias
This change introduces a new `using` alias, `ScopedCallbackHandle`,
inside the `WithCallbacks` class.

This alias defines the return type of `makeScopedCallback`
(a `std::unique_ptr` to a `ScopedEventCallback`), making it easier
for clients to store and reference the handle.

To enable this alias, the `ScopedEventCallback` class definition
was also moved to appear before the `WithCallbacks` class in the header.

Change-Id: Ieb67f75b49797d5998c1c714e15bb40bee29b577