feat(build): Move amc, the Android Meson Configurator tool
This change introduces a new Python-based command-line tool named `amc`
(Android Meson Configurator), located in `tools/toolchain`.

The purpose of this tool is to simplify the configuration and compilation
of Meson-based projects (e.g., QEMU) within the AOSP build environment.
It automates the setup of hermetic toolchains and bridges the gap
between the Meson build system and AOSP's Bazel-based dependencies by
generating `pkg-config` files on the fly.

This new tool requires a Python environment to run. As a result, this
commit also:
* Adds `@rules_python` as a new Bazel dependency to `MODULE.bazel`.
* Configures a default Python 3.10 toolchain.
* Bumps the `aemu` module version from 0.0.1 to 0.0.2.

Note: This is a move of the source from //hardware/generic/goldfish

Bug: 452693757
Change-Id: Id03213fe6be289e9c85616ff5025793e4a6908a6