amc: Add fields to resolve absolute path back to bazel target.
There's a scenario where a field in a custom_target's command uses
an absolute path to some program. That program could've been coming
from one of the pkg-config files we created. So we need to provide
additional fields to give meson's bazel backend the ability to resolve
it back to valid bazel targets.

The two new fields in the .pc file are:
- amc_bazel_target: The bazel target that amc_bin_target is associated
  with.
- amc_bin_target: the absolute path to the binary. This would be the one
  being checked for an exact match.

Bug: 428923963
Test: bazel test //hardware/google/aemu/tools/toolchain/tests/amc:custom_target_test
Change-Id: Ieb52cb18195bd964e4b507d07333b3839a1ed110